Feat: Destination에 Icon 분리
