+++ app/src/main/java/inu/appcenter/intip_android/ui/modifier/ShadowCustom.kt
@@ -1,0 +1,64 @@
+package inu.appcenter.intip_android.ui.modifier
+
+import android.graphics.BlurMaskFilter
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.composed
+import androidx.compose.ui.draw.drawBehind
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.Paint
+import androidx.compose.ui.graphics.drawscope.drawIntoCanvas
+import androidx.compose.ui.graphics.toArgb
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+import inu.appcenter.intip_android.ui.navigate.px
+
+fun Modifier.shadowCustom(
+    color: Color = Color.Black,
+    offsetX: Dp = 0.dp,
+    offsetY: Dp = 0.dp,
+    blurRadius: Dp = 0.dp,
+    shapeRadius: Dp = 0.dp,
+) = composed {
+    val paint: Paint = remember { Paint() }
+    val blurRadiusPx = blurRadius.px(LocalDensity.current)
+    val maskFilter = remember {
+        BlurMaskFilter(blurRadiusPx, BlurMaskFilter.Blur.NORMAL)
+    }
+    drawBehind {
+        drawIntoCanvas { canvas ->
+            val frameworkPaint = paint.asFrameworkPaint()
+            if (blurRadius != 0.dp) {
+                frameworkPaint.maskFilter = maskFilter
+            }
+            frameworkPaint.color = color.toArgb()
+
+            val leftPixel = offsetX.toPx()
+            val topPixel = offsetY.toPx()
+            val rightPixel = size.width + leftPixel
+            val bottomPixel = size.height + topPixel
+
+            if (shapeRadius > 0.dp) {
+                val radiusPx = shapeRadius.toPx()
+                canvas.drawRoundRect(
+                    left = leftPixel,
+                    top = topPixel,
+                    right = rightPixel,
+                    bottom = bottomPixel,
+                    radiusX = radiusPx,
+                    radiusY = radiusPx,
+                    paint = paint,
+                )
+            } else {
+                canvas.drawRect(
+                    left = leftPixel,
+                    top = topPixel,
+                    right = rightPixel,
+                    bottom = bottomPixel,
+                    paint = paint,
+                )
+            }
+        }
+    }
+}
