+++ app/src/main/java/inu/appcenter/intip_android/ui/navigate/AllDestination.kt
@@ -1,10 +1,8 @@
 package inu.appcenter.intip_android.ui.navigate
 
-import inu.appcenter.intip_android.R
-
-sealed class AllDestination(val route: String, val icon: Int, val label: String) {
-    data object Main : AllDestination("main", R.drawable.icon_home_active, "홈")
-    data object Save : AllDestination("save", R.drawable.icon_save, "저장소")
-    data object Write : AllDestination("write", R.drawable.icon_write, "작성")
-    data object MyPage : AllDestination("myPage", R.drawable.icon_profile, "마이페이지")
+sealed class AllDestination(val route: String, val label: String) {
+    data object Main : AllDestination("main", "홈")
+    data object Save : AllDestination("save", "저장소")
+    data object Write : AllDestination("write", "작성")
+    data object MyPage : AllDestination("myPage", "마이페이지")
 }
