+++ app/src/main/java/inu/appcenter/intip_android/ui/navigate/AppBottomBar.kt
@@ -1,12 +1,13 @@
 package inu.appcenter.intip_android.ui.navigate
 
-import android.graphics.BlurMaskFilter
+import androidx.annotation.DrawableRes
+import androidx.compose.foundation.background
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.NavigationBar
@@ -18,34 +19,39 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.composed
-import androidx.compose.ui.draw.drawBehind
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.Paint
-import androidx.compose.ui.graphics.drawscope.drawIntoCanvas
-import androidx.compose.ui.graphics.toArgb
-import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Density
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.currentBackStackEntryAsState
+import androidx.navigation.compose.rememberNavController
+import inu.appcenter.intip_android.R
+import inu.appcenter.intip_android.ui.modifier.shadowCustom
 import inu.appcenter.intip_android.ui.theme.Blue_9CAFE2
 import inu.appcenter.intip_android.ui.theme.Gray_D6D1D5
+import inu.appcenter.intip_android.ui.theme.INTIPTheme
 import inu.appcenter.intip_android.ui.theme.Purple_4071B9
 
 @Composable
 fun AppBottomBar(
     navController: NavHostController
 ) {
-    val screens = listOf(AllDestination.Main, AllDestination.Save, AllDestination.Write, AllDestination.MyPage)
+    val screens = listOf<Pair<AllDestination, @DrawableRes Int>>(
+        AllDestination.Main to R.drawable.icon_home_active ,
+        AllDestination.Save to R.drawable.icon_save ,
+        AllDestination.Write to R.drawable.icon_write ,
+        AllDestination.MyPage to R.drawable.icon_profile ,
+    )
 
+    val shape = RoundedCornerShape(topStart = 20.dp, topEnd = 20.dp)
     NavigationBar(
         containerColor = Color.White,
         modifier = Modifier
             .fillMaxWidth()
-            .height(68.867.dp)
+            .background(color = Color(0xffffffff), shape = shape)
             .shadowCustom(
                 color = Color(0xff190000).copy(alpha = 0.1f),
                 offsetX = (-3).dp,
@@ -56,7 +62,10 @@ fun AppBottomBar(
         val navBackStackEntry by navController.currentBackStackEntryAsState()
         val currentDestination = navBackStackEntry?.destination
 
-        screens.forEach { screen ->
+        screens.forEach { item ->
+            val screen = item.first
+            val icon = item.second
+
             NavigationBarItem(
                 icon = {
                     Column(
@@ -64,13 +73,14 @@ fun AppBottomBar(
                         verticalArrangement = Arrangement.spacedBy(6.89.dp) // 여기서 간격을 조절합니다
                     ) {
                         Icon(
-                            painter = painterResource(screen.icon),
+                            painter = painterResource(icon),
                             contentDescription = screen.route,
                             modifier = Modifier.size(21.dp)
                         )
                         Text(
                             text = screen.label,
-                            style = MaterialTheme.typography.labelSmall // 폰트 변경해야함!!
+                            style = MaterialTheme.typography.labelSmall
+                            //todo: 폰트 변경해야함!!
                         )
                     }
                 },
@@ -99,54 +109,13 @@ fun AppBottomBar(
     }
 }
 
-fun Modifier.shadowCustom(
-    color: Color = Color.Black,
-    offsetX: Dp = 0.dp,
-    offsetY: Dp = 0.dp,
-    blurRadius: Dp = 0.dp,
-    shapeRadius: Dp = 0.dp,
-) = composed {
-    val paint: Paint = remember { Paint() }
-    val blurRadiusPx = blurRadius.px(LocalDensity.current)
-    val maskFilter = remember {
-        BlurMaskFilter(blurRadiusPx, BlurMaskFilter.Blur.NORMAL)
-    }
-    drawBehind {
-        drawIntoCanvas { canvas ->
-            val frameworkPaint = paint.asFrameworkPaint()
-            if (blurRadius != 0.dp) {
-                frameworkPaint.maskFilter = maskFilter
-            }
-            frameworkPaint.color = color.toArgb()
-
-            val leftPixel = offsetX.toPx()
-            val topPixel = offsetY.toPx()
-            val rightPixel = size.width + leftPixel
-            val bottomPixel = size.height + topPixel
-
-            if (shapeRadius > 0.dp) {
-                val radiusPx = shapeRadius.toPx()
-                canvas.drawRoundRect(
-                    left = leftPixel,
-                    top = topPixel,
-                    right = rightPixel,
-                    bottom = bottomPixel,
-                    radiusX = radiusPx,
-                    radiusY = radiusPx,
-                    paint = paint,
-                )
-            } else {
-                canvas.drawRect(
-                    left = leftPixel,
-                    top = topPixel,
-                    right = rightPixel,
-                    bottom = bottomPixel,
-                    paint = paint,
-                )
-            }
-        }
+@Preview
+@Composable
+private fun AppBottomBarPrev() {
+    INTIPTheme {
+        AppBottomBar(navController = rememberNavController())
     }
 }
 
-private fun Dp.px(density: Density): Float =
+fun Dp.px(density: Density): Float =
     with(density) { toPx() }
